+++ src/registry-client.ts
